fix: cmsg_len 标记 unsafe + st_mode macOS 类型兼容
1. cmsg_len() 调用了 unsafe fn CMSG_LEN，自身须标记 unsafe
2. lib.rs:97 中 st.st_mode(S_IFMT) 在 macOS 上类型为 u32(u16)，
   显式 cast 为 u32 消除跨平台类型不匹配

+++ native/src/zygisk_inject/src/lib.rs
@@ -94,7 +94,7 @@ pub extern "C" fn zygisk_companion_entry(socket: i32) {
         let mut entry: Option<CompanionFn> = None;
         let mut st: libc::stat = unsafe { core::mem::zeroed() };
         if unsafe { libc::fstat(fd, &mut st) } == 0
-            && (st.st_mode & libc::S_IFMT) == libc::S_IFREG
+            && (st.st_mode as u32 & libc::S_IFMT as u32) == libc::S_IFREG as u32
         {
             let path_bytes = fd_path_bytes(fd);
             let c_path = unsafe {
